# docs/swagger.yaml
basePath: /
definitions:
  dto.BookListItemResponse:
    properties:
      author:
        type: string
      category:
        type: string
      id:
        type: integer
      image_path:
        type: string
      is_out_of_stock:
        type: boolean
      price:
        type: number
      title:
        type: string
    type: object
  dto.BookResponse:
    properties:
      author:
        type: string
      category:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      price:
        type: number
      publisher:
        type: string
      stock:
        type: integer
      title:
        type: string
    type: object
  dto.BooksListResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/dto.BookListItemResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.ChangePasswordRequest:
    properties:
      current_password:
        maxLength: 100
        minLength: 6
        type: string
      new_password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - current_password
    - new_password
    type: object
  dto.DeleteProfileRequest:
    properties:
      password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - password
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      phone_number:
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.UpdateProfileRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone_number:
        maxLength: 20
        type: string
    required:
    - email
    - name
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple bookstore backend built with Go, Gin, GORM and PostgreSQL.
  title: Bookstore API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate refresh token
      parameters:
      - description: Logout Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate a new access token using a refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh Access Token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register user
      tags:
      - Authentication
  /books:
    get:
      description: Returns a paginated list of books sorted by lowest price first
      parameters:
      - description: Search by title
        in: query
        name: search
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BooksListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: View All Books
      tags:
      - Books
    post:
      consumes:
      - multipart/form-data
      description: Create a new book
      parameters:
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Author
        in: formData
        name: author
        required: true
        type: string
      - description: Publisher
        in: formData
        name: publisher
        type: string
      - description: Category
        in: formData
        name: category
        required: true
        type: string
      - description: Price
        in: formData
        name: price
        required: true
        type: number
      - description: Stock
        in: formData
        name: stock
        required: true
        type: integer
      - description: Book Image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BookResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Add Book
      tags:
      - Books
  /books/{id}:
    delete:
      description: Soft deletes a book. Admin only.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete book
      tags:
      - Books
    get:
      description: Returns a single book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: View One Book
      tags:
      - Books
    put:
      consumes:
      - multipart/form-data
      description: Update an existing book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Author
        in: formData
        name: author
        required: true
        type: string
      - description: Publisher
        in: formData
        name: publisher
        type: string
      - description: Category
        in: formData
        name: category
        required: true
        type: string
      - description: Price
        in: formData
        name: price
        required: true
        type: number
      - description: Stock
        in: formData
        name: stock
        required: true
        type: integer
      - description: Book image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Book
      tags:
      - Books
  /users/change-password:
    put:
      consumes:
      - application/json
      description: Change authenticated user's password
      parameters:
      - description: Passwords
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - Users
  /users/profile:
    delete:
      consumes:
      - application/json
      description: Soft deletes the authenticated user account after verifying password.
        User must not have active orders (pending, confirmed, out_for_delivery).
      parameters:
      - description: Password confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile deleted successfully
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Invalid request / wrong password / active orders exist
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized (missing or invalid token)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (role not allowed)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete My Profile
      tags:
      - Users
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: View Profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update authenticated user's profile
      parameters:
      - description: Profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Edit My Profile
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"